V2/requirements.txt
# Core API
fastapi==0.135.1
uvicorn[standard]==0.41.0
pydantic==2.12.5
pydantic-settings==2.7.1

# Async Kafka & Mongo
aiokafka==0.12.0
motor==3.6.0

# Intelligence & Data
google-genai==1.67.0
yfinance==1.2.0
feedparser==6.0.12
python-dotenv==1.2.2

# Utilities
httpx==0.28.1
loguru==0.7.2
